FastServer/requirements.txt
fastapi
uvicorn
chromadb
transformers
sentence-transformers
rank_bm25
deep-translator
google-generativeai
redis
langgraph
langsmith